変更

+++ resources/js/components/Like.vue
@@ -5,8 +5,18 @@
 </template>
 
 <script>
+import { mapState, mapGetters } from 'vuex'
 export default {
   props: ['postId'],
+  computed: {
+    ...mapState({
+      getCounter: state => state.post.counter,
+      //getLike: state => state.post.check
+    }),
+    ...mapGetters({
+      counter: 'post/getLikes'
+    }) 
+  },
   created() {
     this.feachLike()
   },
@@ -14,30 +24,38 @@ export default {
     return {
       likeOrUnlike: false,
       likeCounter: 0,
-      info:{
-        userId: this.$store.getters['auth/user_id'],
-      }
     }
   },
   methods: {
     async feachLike() {
      const res = await axios.get('/api/likes/' + this.postId)
-     let counter = res.data[0]
-     this.likeCounter = counter
      this.likeOrUnlike = res.data[1]
+     this.likeCounter = res.data[0]
+      //await this.$store.dispatch('post/like_index', this.postId)
+
+
     },
    async toggleLike(){
-      if(this.likeOrUnlike === false){
-        //いいね
-         this.likeOrUnlike = true
-         const res = await axios.put('/api/likes/' + this.postId, this.info)
-         let counter = res.data
-         this.likeCounter = counter   
-      } else {
-        this.likeOrUnlike = false
-        const res = await axios.delete('/api/likes/' + this.postId)
-        let counter = res.data
-        this.likeCounter = counter  
+      if(!this.$store.getters['auth/check']){
+         alert('ログインしてください')
+         this.$router.push('/login')
+         return
+      }else {
+        if(this.likeOrUnlike === false){
+          //いいね
+           this.likeOrUnlike = true
+           const res = await axios.put('/api/likes/' + this.postId)
+            this.likeCounter = res.data
+          // await this.$store.dispatch('post/like', this.postId)
+
+          
+        } else {
+          //いいね解除
+          this.likeOrUnlike = false
+           const res = await axios.delete('/api/likes/' + this.postId)
+           this.likeCounter = res.data
+          //await this.$store.dispatch('post/unlike', this.postId)
+        }
       }
     },
   }
